Clean up how attrs are communicated through scope
Previously, the code that allowed a component to propagate its attrs
down the scope chain relied on several layers of indirection. Primarily,
it used the `self` object to piggyback state through to `bindSelf`,
where it could create and update stable streams that downstream
listeners could subscribe to.

This was necessary because there was no good general purpose hook to
initialize and update the streams in a scope when it changed.

This commit leverages HTMLBars’ new `bindScope` hook, which is always
invoked, even for shadow scopes.

The intent of this hook is for host environments to inspect the state of
the scope and, if they have stable streams, update those streams with
the new value. You can see an example implementation here, where we
propagate the static `attrs` hash into the stable `attrsStream` stream.