Attribute bindings use manualElement
Instead of adding a new system for dealing with attribute bindings, take
Ember’s APIs and convert them into a list of attribute statements
suitable for consumption by HTMLBars. Then, construct an HTMLBars
`manualElement`, passing in the dynamically constructed attributes.

We also pass along any streams passed into the original component,
ensuring that the scope of attributes created in the original template
is the scope used in the manual element.

Also clean up a number of small details.